Created Globalhashicule.
Okay, go with me on this journey: What if there was no such thing as a
null Somethingicule?  What if I didn't need to special-case it all over
the place to deal with globalhashes?  There's enough methods in
Somethingicule to generate the data needed for the latitude and
longitude, right?

With that said, I present Globalhashicule.  This is likely going to be
the most ludicrous, possibly unnecessary refactor I've done yet to this
app, given just how much of Geohash Droid depends on the assumption that
a null Graticule means it's a globalhash.  But hey, I'm already in the
midst of a huge refactor to allow Centicules to exist, so why not clean
up another weird quirk of the codebase along the way?

Let's see if I can pull this off.