Reworked a lot of Info.
The main thrust of this is changing Info so it holds a Somethingicule,
not a Graticule.  This sets off a chain reaction of refactorings,
including changing many calls to point to the Somethingicule instead,
removing any globalhash-specific logic (as Globalhashicule now exists),
etc, and along the way I took the opportunity to add a bunch of
nullability annotations where appropriate.

Note this STILL leaves things that either won't build or won't work
right.  There's just a lot going on with this refactor.