Implement Info-specific wiki data getters in wiki classes.
The wiki getters in WikiUtils depended on the nullishness of Graticule
(that is, whether or not it was a globalhash).  With the existence of
Somethingicule, this made more sense to be moved into Info, and as such,
this commit swaps out the WikiUtils calls and removes them from the
class.