Added code to force status bar lightness as need be.
For some reason, API 35 makes the status bar's icons/text always be in
dark mode while still keeping whatever background is appropriate,
resulting in it being nearly invisible and unreadable in light mode
(that is, white icons on a white background).  This hopefully forces the
icons/text to dark if in light mode (that is, dark icons on a white
background).

It'll also do that for the navigation bars, though that didn't seem to
need it.  Better to be safe, I say. (#89)