Renamed "Send to Maps" to "Share Location".
It's come to my attention it's probably a better idea to name this
feature a bit more generically, given there's other apps and such out
there that can handle a geo: Intent than the Google Maps app.  This
should make things a bit clearer.

At some point I should also check to see if the Radar intent still needs
to be in there...

(#92)