Added transformations to createWikiImage.
This will apply the matrix transform from WikiFragment into the infobox-
stamped image.  I mean, it'll apply it BEFORE it gets stamped with the
infobox.

This hasn't been tested yet, as my test wiki has been giving me major
issues, and this is proving to be a headache to debug.  I'm going to
have to rework this for easier debugging, and I think I'm going to start
with changing it from XML to JSON.

#95