Updated doesWikiPageExist and getWikiVersion to use JSON.
It's come to my attention that XML output in MediaWiki is deprecated,
and the JSON output is just easier to use regardless.  This is the first
step in swapping everything there to use JSON instead.

This is also swapping out using CloseableHttpClient (and friends), as
HttpUrlConnection is definitely simpler and should do the job nicely.

#96