Reimplemented login to use JSON.
This also involves implementing form fields in HttpURLConnection, which
is a pain.  Regardless, at this point, WikiService should be considered
broken until I get the rest of the calls converted.

Also note, at this point the Geohashing wiki is DEFINITELY a new enough
version that we don't need to port the old code path in login anymore.

#96