Reimplemented getWikiPage and putWikiPage to use JSON.
Past this, I think it's just the image part, which of course requires
me to do multipart encoding.  Arrg.

Note to self: When logging in, after auth occurs, you get a NEW session
cookie.  The INITIAL cookie (the one used to request the login token),
while needed for the auth step, is REPLACED by the new cookie with the
authentication.

#96